requirements.txt
gradio>=4.0.0
requests>=2.28.0
Pillow>=9.0.0
huggingface_hub>=0.16.0
diffusers>=0.18.0
transformers>=4.30.0
accelerate>=0.20.0
gitpython>=3.1.30
psutil>=5.9.0
torch>=2.0.0
safetensors>=0.3.1
sentencepiece>=0.1.99
